Create playRound and getPlayerSelection functions
getPlayerSelection returns a string input from a user prompt, and
playRound initiates a round of rocks-paper-scissors. playRound
converts all strings to lowercase, ensures that the player selection
is valid, and compares the player value to the computer value. If the
values are the same, or if the user input is invalid, the user is
prompted to input a selection again.