Refactor game function to allow for game resets
"Play Again?" button appears properly when a game is ended, and it
resets the scores and starts a new game when clicked. Clicking
any of the game buttons before clicking "Play Again?" will not work
because the game button event listener is removed until a new game
is started.

After the first occurence of "Play Again?" getting clicked, the
console reports an Uncaugh DOMException error, because it "Failed
to execute 'removeChild' on 'Node': The Node to be removed is not a
child of this node." This issue needs to be debugged.